@@ -183,17 +183,7 @@ public class SliceBuilderTests extends ESTestCase {
         }
 
         @Override
-        public void setProfile(boolean profile) {
-
-        }
-
-        @Override
-        public boolean isProfile() {
-            return false;
-        }
-
-        @Override
-        public BytesReference cacheKey() throws IOException {
+        public BytesReference cacheKey() {
             return null;
         }
 
