@@ -6,18 +6,19 @@
 
 package org.elasticsearch.xpack.sql.expression.function.scalar.datetime;
 
+import org.elasticsearch.xpack.sql.TestUtils;
 import org.elasticsearch.xpack.sql.expression.Expression;
-import org.elasticsearch.xpack.sql.proto.Mode;
-import org.elasticsearch.xpack.sql.proto.Protocol;
-import org.elasticsearch.xpack.sql.session.Configuration;
 import org.elasticsearch.xpack.sql.tree.AbstractNodeTestCase;
 import org.elasticsearch.xpack.sql.tree.Source;
+
+import java.time.ZoneId;
+import java.time.ZonedDateTime;
+import java.util.Objects;
 
 public class CurrentDateTests extends AbstractNodeTestCase<CurrentDate, Expression> {
 
     public static CurrentDate randomCurrentDate() {
-        return new CurrentDate(Source.EMPTY, new Configuration(randomZone(), Protocol.FETCH_SIZE,
-            Protocol.REQUEST_TIMEOUT, Protocol.PAGE_TIMEOUT, null, Mode.PLAIN, null, null, null));
+        return new CurrentDate(Source.EMPTY, TestUtils.randomConfiguration());
     }
 
     @Override
@@ -32,8 +33,10 @@ public class CurrentDateTests extends AbstractNodeTestCase<CurrentDate, Expressi
 
     @Override
     protected CurrentDate mutate(CurrentDate instance) {
-        return new CurrentDate(instance.source(), new Configuration(randomZone(), Protocol.FETCH_SIZE,
-            Protocol.REQUEST_TIMEOUT, Protocol.PAGE_TIMEOUT, null, Mode.PLAIN, null, null, null));
+        ZonedDateTime now = instance.configuration().now();
+        ZoneId mutatedZoneId = randomValueOtherThanMany(o -> Objects.equals(now.getOffset(), o.getRules().getOffset(now.toInstant())),
+                () -> randomZone());
+        return new CurrentDate(instance.source(), TestUtils.randomConfiguration(mutatedZoneId));
     }
 
     @Override
